Moves playlist classes to feature folder
