--- app/src/main/java/com/kelsos/mbrc/content/playlists/PlaylistRepository.kt
@@ -1,5 +1,0 @@
-package com.kelsos.mbrc.content.playlists
-
-import com.kelsos.mbrc.interfaces.data.Repository
-
-interface PlaylistRepository : Repository<Playlist>
