+++ app/src/main/java/com/kelsos/mbrc/content/sync/LibrarySyncUseCaseImpl.kt
@@ -5,7 +5,7 @@ import com.kelsos.mbrc.content.library.albums.AlbumRepository
 import com.kelsos.mbrc.content.library.artists.ArtistRepository
 import com.kelsos.mbrc.content.library.genres.GenreRepository
 import com.kelsos.mbrc.content.library.tracks.TrackRepository
-import com.kelsos.mbrc.content.playlists.PlaylistRepository
+import com.kelsos.mbrc.features.playlists.repository.PlaylistRepository
 import com.kelsos.mbrc.metrics.SyncMetrics
 import com.kelsos.mbrc.metrics.SyncedData
 import com.kelsos.mbrc.utilities.AppCoroutineDispatchers
