+++ app/src/main/java/com/kelsos/mbrc/features/playlists/PlaylistDto.kt
@@ -1,4 +1,4 @@
-package com.kelsos.mbrc.content.playlists
+package com.kelsos.mbrc.features.playlists
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
