+++ app/src/main/java/com/kelsos/mbrc/features/playlists/PlaylistDtoMapper.kt
@@ -1,5 +1,6 @@
-package com.kelsos.mbrc.content.playlists
+package com.kelsos.mbrc.features.playlists
 
+import com.kelsos.mbrc.features.playlists.data.PlaylistEntity
 import com.kelsos.mbrc.interfaces.data.Mapper
 
 class PlaylistDtoMapper : Mapper<PlaylistDto, PlaylistEntity> {
