+++ app/src/main/java/com/kelsos/mbrc/features/playlists/domain/Playlist.kt
@@ -1,4 +1,4 @@
-package com.kelsos.mbrc.content.playlists
+package com.kelsos.mbrc.features.playlists.domain
 
 data class Playlist(
   val name: String,
