+++ app/src/main/java/com/kelsos/mbrc/features/playlists/repository/PlaylistRepository.kt
@@ -1,0 +1,6 @@
+package com.kelsos.mbrc.features.playlists.repository
+
+import com.kelsos.mbrc.features.playlists.domain.Playlist
+import com.kelsos.mbrc.interfaces.data.Repository
+
+interface PlaylistRepository : Repository<Playlist>
