+++ app/src/main/res/layout/fragment_playlists.xml
@@ -12,7 +12,7 @@
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintStart_toStartOf="parent"
       app:layout_constraintTop_toTopOf="parent"
-      tools:context="com.kelsos.mbrc.ui.navigation.playlists.PlaylistFragment">
+      tools:context="com.kelsos.mbrc.features.playlists.presentation.PlaylistFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/playlists__playlist_list"
